# services/price-service/docs/swagger.yaml
basePath: /internal
definitions:
...
  handlers.GetStorePricesResponse:
    properties:
      prices:
        items:
          $ref: '#/definitions/handlers.StorePrice'
        type: array
      total:
        type: integer
    type: object
...
  handlers.SearchItem:
    properties:
      avgPrice:
        description: Average price across stores
        type: integer
      brand:
        type: string
      category:
        type: string
      chainSlug:
        type: string
      description:
        type: string
      externalId:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      storeCount:
        description: Number of stores with this item
        type: integer
      subcategory:
        type: string
      unit:
        type: string
      unitQuantity:
        type: string
    type: object
  handlers.SearchItemsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.SearchItem'
        type: array
      query:
        type: string
      total:
        type: integer
    type: object
...
  handlers.StorePrice:
    properties:
      anchorPrice:
        type: integer
      brand:
        type: string
      currentPrice:
        type: integer
      discountEnd:
        type: string
      discountPrice:
        type: integer
      discountStart:
        type: string
      inStock:
        type: boolean
      itemExternalId:
        type: string
      itemName:
        type: string
      lastSeenAt:
        type: string
      lowestPrice30d:
        type: integer
      previousPrice:
        type: integer
      priceSignature:
        type: string
      retailerItemId:
        type: string
      unit:
        type: string
      unitPrice:
        type: integer
      unitPriceBaseQuantity:
        type: string
      unitPriceBaseUnit:
        type: string
      unitQuantity:
        type: string
    type: object
info:
  contact: {}
  description: Internal API for price data management, ingestion monitoring, and basket
    optimization.
  title: Price Service API
  version: "1.0"
paths:
...
  /internal/items/search:
    get:
      consumes:
      - application/json
      description: Search for items by name with optional chain filter. Requires minimum
        3 characters.
      parameters:
      - description: Search query (min 3 chars)
        in: query
        minLength: 3
        name: q
        required: true
        type: string
      - description: Filter by chain slug
        in: query
        name: chainSlug
        type: string
      - default: 20
        description: Number of items to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SearchItemsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search items
      tags:
      - items
  /internal/prices/{chainSlug}/{storeId}:
    get:
      consumes:
      - application/json
      description: Returns paginated prices for a specific store in a chain
      parameters:
      - description: Chain slug identifier
        in: path
        name: chainSlug
        required: true
        type: string
      - description: Store ID
        in: path
        name: storeId
        required: true
        type: string
      - default: 100
        description: Number of items to return
        in: query
        maximum: 500
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetStorePricesResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get store prices
      tags:
      - prices
swagger: "2.0"